+++ src/modules/grid.module.ts
@@ -3,7 +3,7 @@ import {CommonModule as AngularCommonModule} from "@angular/common";
 import {CommonModule} from "@anglr/common";
 
 import {GridComponent} from "../components/grid/grid.component";
-import {PageVirtualScrollPagingComponent, ContentVirtualScrollPagingComponent, PreviousNextPagingComponent, LoadMorePagingComponent, BasicPagingComponent, NoPagingComponent, QueryPagingInitializerComponent, NoPagingInitializerComponent} from "../plugins/paging";
+import {PageVirtualScrollPagingComponent, ContentVirtualScrollPagingComponent, PreviousNextPagingComponent, LoadMorePagingComponent, BasicPagingComponent, NoPagingComponent, QueryPagingInitializerComponent, QueryCookiePagingInitializerComponent, NoPagingInitializerComponent} from "../plugins/paging";
 import {SyncDataLoaderComponent, AsyncDataLoaderComponent} from "../plugins/dataLoader";
 import {BasicTableColumnComponent, BasicTableMetadataGathererComponent} from "../components/metadata";
 import {NoMetadataSelectorComponent, AdvancedMetadataSelectorComponent} from "../plugins/metadataSelector";
@@ -31,6 +31,7 @@ import {BasicRowSelectorComponent} from "../plugins/rowSelector";
         PreviousNextPagingComponent,
         NoPagingComponent,
         QueryPagingInitializerComponent,
+        QueryCookiePagingInitializerComponent,
         NoPagingInitializerComponent,
         SyncDataLoaderComponent,
         AsyncDataLoaderComponent,
@@ -56,6 +57,7 @@ import {BasicRowSelectorComponent} from "../plugins/rowSelector";
         PreviousNextPagingComponent,
         NoPagingComponent,
         QueryPagingInitializerComponent,
+        QueryCookiePagingInitializerComponent,
         NoPagingInitializerComponent,
         SyncDataLoaderComponent,
         AsyncDataLoaderComponent,
@@ -84,6 +86,7 @@ import {BasicRowSelectorComponent} from "../plugins/rowSelector";
         PreviousNextPagingComponent,
         NoPagingComponent,
         QueryPagingInitializerComponent,
+        QueryCookiePagingInitializerComponent,
         NoPagingInitializerComponent,
         SyncDataLoaderComponent,
         AsyncDataLoaderComponent,