+++ src/plugins/paging/plugins/pagingInitializer/index.ts
@@ -2,3 +2,5 @@ export * from './no/noPagingInitializer.component';
 export * from './no/noPagingInitializer.interface';
 export * from './query/queryPagingInitializer.component';
 export * from './query/queryPagingInitializer.interface';
+export * from './queryCookie/queryCookiePagingInitializer.component';
+export * from './queryCookie/queryCookiePagingInitializer.interface';